made changes to medicine api

+++ src/main/java/com/project/online_book_store/repository/MedicineRepository.java
@@ -4,7 +4,7 @@ import com.project.online_book_store.entity.Book;
 import com.project.online_book_store.entity.Medicine;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MedicineRepository extends JpaRepository<Medicine, Long> {
+public interface MedicineRepository extends JpaRepository<Medicine, Integer> {
     Medicine findByName(String medicineName);
 
 }
